Add generalized css-to-json-util dev module
